maigrationfile編集

+++ db/migrate/20191118123726_create_reviews.rb
@@ -1,7 +1,7 @@
 class CreateReviews < ActiveRecord::Migration[5.2]
   def change
     create_table :reviews do |t|
-      t.integer :user_id     , null: false
+      t.bigint :user_id     , null: false
       t.integer :fish_id     , null: false
       t.integer :taste_level , null: false
       t.string  :season      , null: false
@@ -10,7 +10,7 @@ class CreateReviews < ActiveRecord::Migration[5.2]
       t.string  :recipe
       t.string  :preservation_period
       t.integer :fish_size   , null: false
-      t.boolean :blood
+      t.boolean :blood       
       t.text    :taste_imp   
       t.string  :picture
